feat: db entry-point
