+++ projects/chit-chat/src/lib/auth/index.ts
@@ -1,0 +1,1 @@
+export * from './public-api';
