+++ projects/chit-chat/src/lib/auth/public-api.ts
@@ -1,0 +1,1 @@
+export * from './services/';
