+++ projects/chit-chat/src/lib/chit-chat.module.ts
@@ -1,45 +1,19 @@
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
-import {
-	InjectionToken,
-	ModuleWithProviders,
-	NgModule,
-} from '@angular/core';
+import { ModuleWithProviders, NgModule } from '@angular/core';
 import {
 	AngularFireModule,
 	FIREBASE_OPTIONS,
 } from '@angular/fire/compat';
 import { AngularFirestoreModule } from '@angular/fire/compat/firestore';
-import { PickerComponent } from '@ctrl/ngx-emoji-mart';
-import { IonicModule } from '@ionic/angular';
-import { AuthService, UserService } from './services';
-
-export interface LibConfig {
-	firebaseConfig: {
-		apiKey: string;
-		authDomain: string;
-		// databaseURL: string;
-		projectId: string;
-		storageBucket: string;
-		messagingSenderId: string;
-		appId: string;
-		measurementId: string;
-	};
-}
-
-export const LibConfigService = new InjectionToken<LibConfig>(
-	'LibConfig'
-);
+import { AuthService } from 'chit-chat/src/lib/auth';
+import {
+	LibConfig,
+	LibConfigService,
+} from 'chit-chat/src/lib/lib-config';
 
 @NgModule({
-	imports: [
-		CommonModule,
-		HttpClientModule,
-		IonicModule,
-		PickerComponent,
-		AngularFireModule,
-		AngularFirestoreModule,
-	],
+	declarations: [],
+	imports: [CommonModule, AngularFireModule, AngularFirestoreModule],
 })
 export class ChitChatModule {
 	static forRoot(
@@ -49,7 +23,6 @@ export class ChitChatModule {
 			ngModule: ChitChatModule,
 			providers: [
 				AuthService,
-				UserService,
 				{
 					provide: FIREBASE_OPTIONS,
 					useValue: config.firebaseConfig,
