+++ projects/chit-chat/src/lib/components/conversation-list/src/conversation-list.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { AuthService } from 'chit-chat/src/lib/auth';
 
 @Component({
 	selector: 'chit-chat-conversation-list',
@@ -6,5 +7,7 @@ import { Component } from '@angular/core';
 	styleUrls: ['./conversation-list.component.scss'],
 })
 export class ConversationListComponent {
-	constructor() {}
+	constructor(private auth: AuthService) {
+		setTimeout(() => console.log(this.auth.getCurrentUser()), 5000);
+	}
 }
