+++ projects/chit-chat/src/lib/database-utils/index.ts
@@ -1,0 +1,1 @@
+export * from './public-api';
