+++ projects/chit-chat/src/lib/database-utils/interfaces/map-result.interface.ts
