+++ projects/chit-chat/src/lib/database-utils/models/index.ts
@@ -1,4 +1,3 @@
-export * from './chat-user.model';
 export * from './permission.model';
 export * from './user-role.model';
 export * from './user.model';
