+++ projects/chit-chat/src/lib/database-utils/models/user-role.model.ts
@@ -1,6 +1,5 @@
 import { DtoPermission, DtoUserRole } from '../dto';
 import { MapResult } from '../interfaces';
-
 import { Permission } from './permission.model';
 
 export class UserRole implements DtoUserRole {
