+++ projects/chit-chat/src/lib/database-utils/ng-package.json
@@ -1,0 +1,6 @@
+{
+	"$schema": "../../../../../node_modules/ng-packagr/ng-package.schema.json",
+	"lib": {
+		"entryFile": "public-api.ts"
+	}
+}
