+++ projects/chit-chat/src/lib/lib-config/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './lib-config.interface';
