+++ projects/chit-chat/src/lib/lib-config/interfaces/lib-config.interface.ts
@@ -1,0 +1,12 @@
+export interface LibConfig {
+	firebaseConfig: {
+		apiKey: string;
+		authDomain: string;
+		// databaseURL: string;
+		projectId: string;
+		storageBucket: string;
+		messagingSenderId: string;
+		appId: string;
+		measurementId: string;
+	};
+}
