+++ projects/chit-chat/src/lib/lib-config/public-api.ts
@@ -1,0 +1,2 @@
+export * from './interfaces/';
+export * from './services/';
