+++ projects/chit-chat/src/lib/lib-config/services/index.ts
@@ -1,0 +1,1 @@
+export * from './lib-config.service';
