+++ projects/chit-chat/src/lib/lib-config/services/lib-config.service.ts
@@ -1,0 +1,6 @@
+import { InjectionToken } from '@angular/core';
+import { LibConfig } from '../interfaces';
+
+export const LibConfigService = new InjectionToken<LibConfig>(
+	'LibConfig'
+);
