--- projects/chit-chat/src/lib/styles.scss
@@ -1,8 +1,0 @@
-// ::ng-deep .emoji-container{
-//     display: relative;
-
-//     .emoji-mart-bar{
-//         position: absolute;
-//         bottom: 0;
-//     }
-// }
